requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
Deprecated==1.2.13
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.0
fastapi==0.70.0
fastapi-mail==1.0.2
h11==0.12.0
httpcore==0.13.7
httptools==0.3.0
httpx==0.19.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pymongo==4.0.1
pyparsing==3.0.6
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.0.2
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.16.0
typing_extensions==4.0.1
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
